+++ src/main/scala/ch/openolitor/core/data/parsers/AboParser.scala
@@ -60,7 +60,7 @@ object AboParser extends EntityParser {
         //calculate count
         val anzahlAbwesenheiten = parseTreeMap(row.value[String](indexAnzahlAbwesenheiten))(identity, _.toInt)
         val anzahlLieferungen = parseTreeMap(row.value[String](lieferungenIndex))(identity, _.toInt)
-        val anzahlEinsaetze = parseTreeMap(row.value[String](einsaetzeIndex))(identity, _.toInt)
+        val anzahlEinsaetze = parseTreeMap(row.value[String](einsaetzeIndex))(identity, BigDecimal(_))
 
         val erstelldat = row.value[DateTime](indexErstelldat)
         val ersteller = PersonId(row.value[Long](indexErsteller))
